requirements.txt
playwright
beautifulsoup4
pandas
psycopg2-binary
python-dotenv
